+++ src/main/java/backend/wal/wal/item/domain/aggregate/Item.java
@@ -1,4 +1,4 @@
-package backend.wal.wal.nextwal.domain.aggregate;
+package backend.wal.wal.item.domain.aggregate;
 
 import lombok.AccessLevel;
 import lombok.Builder;
