+++ src/main/java/backend/wal/wal/item/domain/repository/ItemRepository.java
@@ -1,7 +1,7 @@
-package backend.wal.wal.nextwal.domain.repository;
+package backend.wal.wal.item.domain.repository;
 
 import backend.wal.wal.common.domain.WalCategoryType;
-import backend.wal.wal.nextwal.domain.aggregate.Item;
+import backend.wal.wal.item.domain.aggregate.Item;
 
 import org.springframework.data.jpa.repository.JpaRepository;
