+++ src/main/java/backend/wal/wal/nextwal/domain/aggregate/NextWal.java
@@ -2,6 +2,7 @@ package backend.wal.wal.nextwal.domain.aggregate;
 
 import backend.wal.wal.common.domain.WalCategoryType;
 
+import backend.wal.wal.item.domain.aggregate.Item;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
